Restored codegen -extended 64K code segment option
Last remaining issue was overflow on loading (see HopperSystem.hs)

+++ Runtimes/Windows/HopperNET/HopperSystem.cs
@@ -71,7 +71,7 @@ namespace HopperNET
 #endif
             byte[] code = new byte[fileSize];
             HopperFile binaryFile = HopperFile.Open(ProgramPath);
-            ushort iCode = 0;
+            uint iCode = 0;
             for (; ; )
             {
                 byte b = binaryFile.Read();
